refactor(core): 将 Redis 封装迁出 helper 至 app.core.redis，消除 core/cache→helper 反向依赖
RedisHelper/AsyncRedisHelper 是依赖 settings 的缓存后端基础设施，却位于 app/helper/redis.py，
导致 core/cache.py 顶层反向依赖 helper 层（架构审计 P1：core 应为零 helper 依赖的稳定底座）。

改动：
- git mv app/helper/redis.py -> app/core/redis.py（内容不变；仅依赖 core.config 与 utils，
  方向正确；ConfigReloadMixin 的 config.updated 订阅在新 qualname 下经验证正常工作）。
- app/helper/redis.py 改为 re-export 兼容垫片（RedisHelper / AsyncRedisHelper / serialize /
  deserialize），保留历史导入路径——含社区插件 p115strmhelper（utils/tree、core/cache），
  插件无需任何改动即可继续工作。
- 3 处内部 importer 改为 from app.core.redis：core/cache、startup/modules_initializer、
  modules/redis。
- 新增 test_core_redis_relocation.py：垫片同类（含插件路径）/ serialize 垫片可用 /
  core/cache 不再引用 helper.redis / core.redis 无 helper 依赖。

效果：消除 core/cache→helper.redis top-level 反向边；行为零变化；插件契约经垫片完整保留。
